pyproject.toml
[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E","F","I","B","UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[project]
name = "telebot-starter"
version = "0.1.0"
description = "Telegram bot with weather, quote, calendar (stub), and meme commands"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot==21.6",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.2", "pre-commit>=4.0.1", "black>=24.8.0", "ruff>=0.6.9"]